+++ frameworks/js/napi/session/include/napi_avcast_controller_callback.h
@@ -45,7 +45,7 @@ public:
     void OnMediaItemChange(const AVQueueItem& avQueueItem) override;
     void OnVolumeChange(const int32_t volume) override;
     void OnLoopModeChange(const int32_t loopMode) override;
-    void OnPlaySpeedChange(const int32_t playSpeed) override;
+    void OnPlaySpeedChange(const double playSpeed) override;
     void OnPositionChange(const int32_t position) override;
     void OnVideoSizeChange(const int32_t width, const int32_t height) override;
     void OnError(const int32_t errorCode, const std::string& errorMsg) override;
